make wsgi_eager.py the default (cherry picked from commit e1d0d12bf7951d2c17dbc41b86ab5b7b71c0a3f1)
